requirements.txt
iso639-lang==2.1.0
request==2.28.2
